add an Important to VPCAccessPolicy (#54)
VPCAccessPolicy is expecting and object, even an empty one.  Not passing an object will result in an error.  Eric Johnson (ericdj@amzon.com) helped me out with this solution.   I also submitted feedback, but think either a Note or Important is a good thing to add until the document could be otherwise updated.

Co-authored-by: John Walsh <john.walsh@braviant.com>
Co-authored-by: Mark Hirayama <34664502+mark-hirayama@users.noreply.github.com>