Update serverless-policy-templates.md (#102)
Edit the verbiage to avoid the confusion that the "S3ReadPolicy" is not to give permission to the objects in S3, but rather, to give permission to read the objects in S3. The original verbiage "Gives read-only permission to objects in an Amazon Simple Storage Service (Amazon S3) bucket.", could be misleading.  Same with the other S3 policy description.