Update monitor-app-insights.md
The Application Insights application should use an intrinsic reference to the resource group rather than duplicating the string concatenation used to build the resource group name.  This has two benefits:  First, it creates an implicit dependency, removing the need for a `DependsOn`.  Second, if the name of the resource group changes, then the Application Insights will automatically be updated to reference the correct resource group.